Move exception and result filters in base
